requirements.txt
yfinance
pandas
numpy
matplotlib
scipy
tqdm
python-dotenv
pydantic>1.10